requirements.txt
django==4.1.7
djangorestframework==3.14.0
drf_yasg==1.21.5
autopep8
flake8
